// generated/src/app/api-models/medi-cal-php-address-mstr.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MediCalPhpAddressMstr {
  seqPhpAddressId: number;
  seqPrediId: number;
  monthOfEligibility: Date;
  holdStatus: string;
  enrollFlag: string;
  projectNumber: string;
  countyId: string;
  countyCode: string;
  aidCode: string;
  caseNumber: string;
  fbu: string;
  personNumber: string;
  lastName: string;
  firstName: string;
  middleInitial: string;
  gender: string;
  dateOfBirth: Date;
  medicareNo: string;
  district: string;
  currentRestrict: string;
  currentMedicareStatus: string;
  currentOhcStatus: string;
  currentEligStatus: string;
  currentPhpStatus: string;
  janMedicareStatus: string;
  janOhcStatus: string;
  janEligStatus: string;
  janPhpStatus: string;
  febMedicareStatus: string;
  febOhcStatus: string;
  febEligStatus: string;
  febPhpStatus: string;
  marMedicareStatus: string;
  marOhcStatus: string;
  marEligStatus: string;
  marPhpStatus: string;
  aprMedicareStatus: string;
  aprOhcStatus: string;
  aprEligStatus: string;
  aprPhpStatus: string;
  mayMedicareStatus: string;
  mayOhcStatus: string;
  mayEligStatus: string;
  mayPhpStatus: string;
  junMedicareStatus: string;
  junOhcStatus: string;
  junEligStatus: string;
  junPhpStatus: string;
  julMedicareStatus: string;
  julOhcStatus: string;
  julEligStatus: string;
  julPhpStatus: string;
  augMedicareStatus: string;
  augOhcStatus: string;
  augEligStatus: string;
  augPhpStatus: string;
  sepMedicareStatus: string;
  sepOhcStatus: string;
  sepEligStatus: string;
  sepPhpStatus: string;
  octMedicareStatus: string;
  octOhcStatus: string;
  octEligStatus: string;
  octPhpStatus: string;
  novMedicareStatus: string;
  novOhcStatus: string;
  novEligStatus: string;
  novPhpStatus: string;
  decMedicareStatus: string;
  decOhcStatus: string;
  decEligStatus: string;
  decPhpStatus: string;
  careOfAddress: string;
  streetAddress: string;
  cityAndState: string;
  zipCode: string;
  medsId: string;
  issueDate: Date;
  userDefined1: string;
  userDefined2: string;
  eligUserDefined1: string;
  groupId: string;
  planCode: string;
  country: string;
}